implements saving of entered data in the form to the component state
