+++ src/components/offer-details/offer-details.tsx
@@ -1,19 +1,11 @@
-import { Fragment } from 'react';
 import { TOffer } from '../../types/offer-type';
 import { TReviews } from '../../types/reviews-type';
+import ReviewForm from '../review-form/review-form';
 
 type TOfferPage = {
   offer: TOffer;
   reviews: TReviews[];
 }
-
-const ratingValue = {
-  1: 'terribly',
-  2: 'badly',
-  3: 'not bad',
-  4: 'good',
-  5: 'perfect'
-};
 
 function OffersDetails({offer, reviews}: TOfferPage): JSX.Element {
   const {title, images, type, goods, price, bedrooms, rating, isPremium, maxAdults, host, description} = offer;
@@ -130,55 +122,7 @@ function OffersDetails({offer, reviews}: TOfferPage): JSX.Element {
                   </div>
                 </li>
               </ul>
-              <form className="reviews__form form" action="#" method="post">
-                <label className="reviews__label form__label" htmlFor="review">
-              Your review
-                </label>
-                <div className="reviews__rating-form form__rating">
-                  {Object.entries(ratingValue).map(([key, value]) => (
-                    <Fragment key={key}>
-                      <input
-                        className="form__rating-input visually-hidden"
-                        name="rating"
-                        defaultValue={key}
-                        id={`${key}-stars`}
-                        type="radio"
-                      />
-                      <label
-                        htmlFor={`${key}-stars`}
-                        className="reviews__rating-label form__rating-label"
-                        title={value}
-                      >
-                        <svg className="form__star-image" width={37} height={33}>
-                          <use xlinkHref="#icon-star" />
-                        </svg>
-                      </label>
-                    </Fragment>
-                  ))}
-                </div>
-                <textarea
-                  className="reviews__textarea form__textarea"
-                  id="review"
-                  name="review"
-                  placeholder="Tell how was your stay, what you like and what can be improved"
-                  defaultValue={''}
-                />
-                <div className="reviews__button-wrapper">
-                  <p className="reviews__help">
-                To submit review please make sure to set{' '}
-                    <span className="reviews__star">rating</span> and describe your
-                stay with at least{' '}
-                    <b className="reviews__text-amount">50 characters</b>.
-                  </p>
-                  <button
-                    className="reviews__submit form__submit button"
-                    type="submit"
-                    disabled
-                  >
-                Submit
-                  </button>
-                </div>
-              </form>
+              <ReviewForm />
             </section>
           ))}
